Move connect and login into m.c.bones.helpers.
